backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0          # ← needed for SQLite demo mode (DATABASE_URL=sqlite+aiosqlite:///)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
boto3==1.34.100
python-dotenv==1.0.1
httpx==0.27.0
razorpay==1.4.1
